requirements.txt
streamlit==1.17.0
torch==2.1.0
torchvision==0.16.0
Pillow==9.5.0
gdown==4.6.0
